loginObservable accepts redirect_uri as a parameter
Previously, the redirect_uri for the loginObservable flow was hardcoded to be Jonas's episphere/epibox Observable notebook. That would mean that no one else would be able to integrate it with their Observable notebook directly. Making it a parameter so that anyone with an Observable notebook might use it in theory (although they can't really, because they can't set the redirect_URI on the Box app without being given access to edit it.)